fix: 상금 합 오버플로 발생 버그 수정

+++ src/main/java/domain/Rewards.java
@@ -14,7 +14,7 @@ public enum Rewards {
         this.prize = prize;
     }
 
-    public int getPrize() {
+    public long getPrize() {
         return this.prize;
     }
 }
